build: pin jsonschema to ^4.17.3 to avoid installing 4.18.x
Because some dependencies of jsonschema now require the rust toolchain (see 1 and 2), building
charms that depend on SDI (and therefore pull jsonschema) fail when building because none
of them have rustc available when pip installing the charm's python dependencies.

Reasoning:
jsonschema 4.18.x introduced a dependency on rdps-py which at the same time introduced
a dependency on maturin, a python package that requires the rustc toolchain to be
built and installed from source. jsonschema 4.17.x does not have that dependency.

Links:
[1]python-jsonschema/jsonschema#1117
[2] python-jsonschema/jsonschema#1114

Partially fixes: canonical/bundle-kubeflow#648